+++ .gitignore
@@ -117,4 +117,5 @@ summary.html
 # common labels template result files of make target module-version
 config/default/kustomization.yaml-e
 
-internal/manifest/test/merged-istio-operator.yaml
+# ignore generated files for istiooperator tests
+internal/istiooperator/test/merged-istio-operator.yaml
