+++ api/v1alpha2/resources_test.go
@@ -77,11 +77,11 @@ var _ = Describe("GetProxyResources", func() {
 
 	It("should be able to get resources from real istio operator template when IstioCR has no overrides", func() {
 		//given
-		manifest, err := os.ReadFile("../../internal/istiooperator/istio-operator.yaml")
+		istioOperator, err := os.ReadFile("../../internal/istiooperator/istio-operator.yaml")
 		Expect(err).ShouldNot(HaveOccurred())
 
 		iop := iopv1alpha1.IstioOperator{}
-		err = yaml.Unmarshal(manifest, &iop)
+		err = yaml.Unmarshal(istioOperator, &iop)
 		Expect(err).ShouldNot(HaveOccurred())
 
 		istioCR := v1alpha2.Istio{}
